+++ src/models/transaction.rs
@@ -1,7 +1,7 @@
 // models.rs
 use chrono::{DateTime, Utc, NaiveDate};
 use serde::{Deserialize, Serialize};
-use sqlx::FromRow;
+use sqlx::{FromRow, types::BigDecimal};
 use uuid::Uuid;
 
 #[derive(Debug, FromRow, Serialize, Deserialize, Clone)]
@@ -10,7 +10,7 @@ pub struct Transaction {
     pub description: Option<String>,
     pub debit_account_id: Uuid,
     pub credit_account_id: Uuid,
-    pub amount: f64,
+    pub amount: BigDecimal,
     pub date: NaiveDate,
     pub user_id: Uuid,
     pub created_at: DateTime<Utc>,
@@ -21,5 +21,5 @@ pub struct CreateTransaction {
     pub description: Option<String>,
     pub debit_account_id: Uuid,
     pub credit_account_id: Uuid,
-    pub amount: f64,
+    pub amount: BigDecimal,
 }
